requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.3.4
billiard==3.3.0.23
cairocffi==0.7.2
CairoSVG==1.0.22
celery==3.1.23
cffi==1.6.0
cssselect==0.9.2
django-celery==3.1.17
django-paypal==0.3.2
flower==0.9.1
html5lib==0.9999999
kombu==3.0.35
lxml==3.6.0
Pillow==3.2.0
pycparser==2.14
Pyphen==0.9.4
pytz==2016.4
requests==2.10.0
six==1.10.0
tinycss==0.3
tornado==4.2
WeasyPrint==0.28
Django==1.9.7
